feat: add host macro
This migrates the host macro over to the repo
from the example repo
Current hurdle is there is a hard coded assumption
on the name of the npm workspace in which webpack will
be loaded from. That needs to be refactored.
In addition, this repo needs to specify webpack as a dependency
which is still in development